+++ app/templates/_beneficjenci_rows.html
@@ -3,10 +3,10 @@
   <td>{{ b.imie }}</td>
   <td>{{ b.wojewodztwo }}</td>
   <td>
-    <a href="{{ url_for('edytuj_beneficjenta', beneficjent_id=b.id) }}" class="btn btn-sm btn-primary" aria-label="Edytuj" data-bs-toggle="tooltip" title="Edytuj">
+    <a href="{{ url_for('sessions.edytuj_beneficjenta', beneficjent_id=b.id) }}" class="btn btn-sm btn-primary" aria-label="Edytuj" data-bs-toggle="tooltip" title="Edytuj">
       <i class="bi bi-pencil" aria-hidden="true"></i>
     </a>
-    <form method="post" action="{{ url_for('usun_beneficjenta', beneficjent_id=b.id) }}" style="display:inline;">
+    <form method="post" action="{{ url_for('sessions.usun_beneficjenta', beneficjent_id=b.id) }}" style="display:inline;">
       {{ delete_form.csrf_token }}
       <button type="submit" class="btn btn-sm btn-danger" onclick="return confirm('Na pewno chcesz usunąć?');" aria-label="Usuń" data-bs-toggle="tooltip" title="Usuń">
         <i class="bi bi-trash" aria-hidden="true"></i>
