+++ app/templates/login.html
@@ -6,7 +6,7 @@
 <p class="mb-4 text-muted">Podaj swój adres email i hasło, aby się zalogować.</p>
 <div class="row justify-content-center text-center">
   <div class="col-12 col-md-6 d-flex justify-content-center">
-    <form method="post" action="{{ url_for('login') }}" class="text-start w-100" style="max-width: 400px;">
+    <form method="post" action="{{ url_for('auth.login') }}" class="text-start w-100" style="max-width: 400px;">
       {{ form.hidden_tag() }}
       {{ render_field(form.email, "adres@email.pl") }}
       {{ render_password_field(form.password, "Hasło", "current-password") }}
@@ -18,8 +18,8 @@
   </div>
 </div>
 <p class="mt-3">Nie masz konta?
-  <a href="{{ url_for('register') }}">Zarejestruj się</a>
+  <a href="{{ url_for('auth.register') }}">Zarejestruj się</a>
 </p>
-<p class="mt-2"><a href="{{ url_for('reset_password_request') }}">Zapomniałeś hasła?</a></p>
+<p class="mt-2"><a href="{{ url_for('auth.reset_password_request') }}">Zapomniałeś hasła?</a></p>
 {% endblock %}
