+++ app/templates/zajecia_calendar.html
@@ -11,7 +11,7 @@ document.addEventListener('DOMContentLoaded', function() {
   var calendarEl = document.getElementById('calendar');
   var calendar = new FullCalendar.Calendar(calendarEl, {
     initialView: 'dayGridMonth',
-    events: '{{ url_for("api_zajecia") }}'
+    events: '{{ url_for("sessions.api_zajecia") }}'
   });
   calendar.render();
 });
